Rename to fibro and publish on pypi

+++ hatch_build.py
@@ -8,7 +8,7 @@ from hatchling.builders.hooks.plugin.interface import BuildHookInterface
 # Bug is described in https://github.com/pypa/hatch/issues/1197
 
 
-symlink = Path(__file__).parent / 'filebrowser/helix/contrib/themes'
+symlink = Path(__file__).parent / 'fibro/helix/contrib/themes'
 target = Path('../runtime/themes')
 
 
@@ -16,9 +16,6 @@ class CustomBuildHook(BuildHookInterface):
 
     def initialize(self, *args, **kwargs):
         super().initialize(*args, **kwargs)
-        assert symlink.is_symlink() and symlink.readlink() == target
-        symlink.unlink()
-
-    def finalize(self, *args, **kwargs):
-        symlink.symlink_to(target)
-        super().finalize(*args, **kwargs)
+        if symlink.exists():
+            assert symlink.is_symlink() and symlink.readlink() == target
+            symlink.unlink()
